# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/events": {
      "get": {
        "operationId": "EventsController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "List of events"
          }
        },
        "summary": "List all events",
        "tags": [
          "Events"
        ]
      },
      "post": {
        "operationId": "EventsController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created"
          },
          "400": {
            "description": "Validation / business rule error"
          }
        },
        "summary": "Create a new event",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Get event by ID",
        "tags": [
          "Events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated"
          }
        },
        "summary": "Update event",
        "tags": [
          "Events"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted"
          }
        },
        "summary": "Delete event",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/organizer/{organizerId}": {
      "get": {
        "operationId": "EventsController_findByOrganizer",
        "parameters": [
          {
            "name": "organizerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get events by organizer ID",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/{id}/registrations-count": {
      "get": {
        "operationId": "EventsController_registrationsCount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get registration count for an event",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/registrations": {
      "get": {
        "operationId": "RegistrationsController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all registrations",
        "tags": [
          "Registrations"
        ]
      },
      "post": {
        "operationId": "RegistrationsController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration created"
          },
          "400": {
            "description": "Already registered / capacity full"
          }
        },
        "summary": "Register for an event",
        "tags": [
          "Registrations"
        ]
      }
    },
    "/api/registrations/event/{eventId}": {
      "get": {
        "operationId": "RegistrationsController_findByEvent",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get registrations for an event",
        "tags": [
          "Registrations"
        ]
      }
    },
    "/api/registrations/user/{userId}": {
      "get": {
        "operationId": "RegistrationsController_findByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get registrations for a user",
        "tags": [
          "Registrations"
        ]
      }
    },
    "/api/registrations/{id}": {
      "delete": {
        "operationId": "RegistrationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Registration cancelled"
          }
        },
        "summary": "Cancel registration",
        "tags": [
          "Registrations"
        ]
      }
    },
...
    "/api/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List notifications (optionally filter by userId)",
        "tags": [
          "Notifications"
        ]
      },
      "post": {
        "operationId": "NotificationsController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification created"
          }
        },
        "summary": "Create a notification",
        "tags": [
          "Notifications"
        ]
      }
    }
  },
  "info": {
    "title": "EMCP API",
    "description": "Event Management & Coordination Platform REST API. Role is passed via the x-user-role header. User ID is passed via the x-user-id header for user-specific operations.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Tech Innovators Forum 2026"
          },
          "description": {
            "type": "string",
            "example": "A day of discovery and demos."
          },
          "imageUrl": {
            "type": "string",
            "example": ""
          },
          "eventDate": {
            "type": "string",
            "example": "2026-05-15"
          },
          "startTime": {
            "type": "string",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "example": "12:00"
          },
          "capacity": {
            "type": "number",
            "example": 250
          },
          "organizerId": {
            "type": "string",
            "example": "u3"
          },
          "venueId": {
            "type": "string",
            "example": "v1"
          }
        },
        "required": [
          "title",
          "eventDate",
          "startTime",
          "endTime",
          "capacity",
          "organizerId",
          "venueId"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "eventDate": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "organizerId": {
            "type": "string"
          },
          "venueId": {
            "type": "string"
          }
        }
      },
      "CreateRegistrationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "u4"
          },
          "eventId": {
            "type": "string",
            "example": "e1"
          }
        },
        "required": [
          "userId",
          "eventId"
        ]
      },
...
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "u5"
          },
          "title": {
            "type": "string",
            "example": "New task assigned!"
          },
          "message": {
            "type": "string",
            "example": "You have been added to the support team."
          },
          "type": {
            "type": "string",
            "example": "info"
          }
        },
        "required": [
          "userId",
          "title",
          "message"
        ]
      }
    }
  }
}